requirements.txt
# Core API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database & Caching
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.11.0
redis>=5.0.0

# Data Science & Market Data
pandas>=2.0.0
numpy>=1.24.0
pandas-market-calendars>=4.0.0
ccxt>=4.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.24.0
loguru>=0.7.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# --- Research & Visualization (Phase 1, Point 9) ---
jupyterlab>=4.0.0
mplfinance
plotly>=5.15.0
ipykernel>=6.25.0
matplotlib>=3.7.0